stdlib: Use long long int in stdlib/tst-llabs
And adjust for GNU style.

+++ stdlib/tst-llabs.c
@@ -25,29 +25,29 @@
 
 static int do_test (void)
 {
-  long i;
+  long long int i;
 
-  TEST_COMPARE(llabs(LLONG_MAX), LLONG_MAX);
-  TEST_COMPARE(llabs(LLONG_MIN + 1), LLONG_MAX);
-  TEST_COMPARE(llabs(0x00000000ffffffffL), 0x00000000ffffffffL);
-  TEST_COMPARE(llabs(0x0000000100000000L), 0x0000000100000000L);
-  TEST_COMPARE(llabs(0x80000000ffffffffL), 0x7fffffff00000001L);
-  TEST_COMPARE(llabs(0x8000000100000000L), 0x7fffffff00000000L);
-  TEST_COMPARE(llabs(-1), 1);
-  TEST_COMPARE(llabs(0), 0);
-  TEST_COMPARE(llabs(1), 1);
+  TEST_COMPARE (llabs (LLONG_MAX), LLONG_MAX);
+  TEST_COMPARE (llabs (LLONG_MIN + 1), LLONG_MAX);
+  TEST_COMPARE (llabs (0x00000000ffffffffL), 0x00000000ffffffffL);
+  TEST_COMPARE (llabs (0x0000000100000000L), 0x0000000100000000L);
+  TEST_COMPARE (llabs (0x80000000ffffffffL), 0x7fffffff00000001L);
+  TEST_COMPARE (llabs (0x8000000100000000L), 0x7fffffff00000000L);
+  TEST_COMPARE (llabs (-1), 1);
+  TEST_COMPARE (llabs (0), 0);
+  TEST_COMPARE (llabs (1), 1);
 
   for (i = LLONG_MIN + 1; i < LLONG_MIN + INT_MAX; i += LARGE_PRIME)
-    TEST_COMPARE(llabs(i), -i);
+    TEST_COMPARE (llabs (i), -i);
 
   for (i = LLONG_MAX - INT_MAX; i < LLONG_MAX - LARGE_PRIME; i += LARGE_PRIME)
-    TEST_COMPARE(llabs(i), i);
+    TEST_COMPARE (llabs (i), i);
 
   for (i = INT_MIN + 1; i < 0; i += LARGE_PRIME)
-    TEST_COMPARE(llabs(i), -i);
+    TEST_COMPARE (llabs (i), -i);
 
   for (i = 0; i < INT_MAX; i += LARGE_PRIME)
-    TEST_COMPARE(llabs(i), i);
+    TEST_COMPARE (llabs (i), i);
 
   return EXIT_SUCCESS;
 }
